named temp_logger, also class name
